Find c++ proposal status on github and some website Split JS comment up by author and detect author Update to new schema

+++ js/main.py
@@ -1,7 +1,7 @@
 import logging
 import os
 from pathlib import Path
-from typing import Dict
+from typing import Dict, List
 
 from dotenv import load_dotenv
 from pydantic import TypeAdapter
@@ -26,33 +26,37 @@ def main():
     proposal_v2_path = Path("./js/output/proposals_v2.json")
     meeting_mapping_path = Path("./js/output/meeting_notes_mapping.json")
     db_path = Path("./js/output/js_proposals.sqlite3")
+    project_id = 3
 
-    project_id = 1
     adapter1 = TypeAdapter(Dict[str, proposal_stages.ProposalV1])
     adapter2 = TypeAdapter(Dict[str, proposal_revisions.ProposalV2])
-    adapter_meetings_map = TypeAdapter(Dict[str, MeetingNote])
+    adapter_meetings_map = TypeAdapter(Dict[str, List[MeetingNote]])
 
-    logging.info("Computing stage history")
-    proposals_v1: Dict[str, proposal_stages.ProposalV1] = (
-        proposal_stages.compute_stage_history(project_id=project_id)
-    )
-    json_data = adapter1.dump_json(proposals_v1, indent=2)
-    proposal_v1_path.write_bytes(json_data)
+    if not proposal_v2_path.exists():
+        logging.info("Computing stage history")
+        proposals_v1: Dict[str, proposal_stages.ProposalV1] = (
+            proposal_stages.compute_stage_history()
+        )
+        json_data = adapter1.dump_json(proposals_v1, indent=2)
+        proposal_v1_path.write_bytes(json_data)
 
-    logging.info("Fetching spec revisions")
-    proposals_v1 = adapter1.validate_json(proposal_v1_path.read_bytes())
-    proposals_v2 = proposal_revisions.fetch_revisions(proposals_v1)
-    json_data = adapter2.dump_json(proposals_v2, indent=2)
-    proposal_v2_path.write_bytes(json_data)
+    if not proposal_v2_path.exists():
+        logging.info("Fetching spec revisions")
+        proposals_v1 = adapter1.validate_json(proposal_v1_path.read_bytes())
+        proposals_v2 = proposal_revisions.fetch_revisions(proposals_v1)
+        json_data = adapter2.dump_json(proposals_v2, indent=2)
+        proposal_v2_path.write_bytes(json_data)
 
-    logging.info("Finding meeting notes")
-    mapping = build_mapping()
-    meeting_mapping_path.write_bytes(adapter_meetings_map.dump_json(mapping, indent=2))
+    if not meeting_mapping_path.exists():
+        logging.info("Finding meeting notes")
+        mapping = build_mapping()
+        meeting_mapping_path.write_bytes(adapter_meetings_map.dump_json(mapping, indent=2))
 
-    logging.info("Saving to database")
-    proposals_v2 = adapter2.validate_json(proposal_v2_path.read_bytes())
-    mapping = adapter_meetings_map.validate_json(meeting_mapping_path.read_bytes())
-    save_to_db(db_path, proposals_v2, mapping, project_id)
+    if not db_path.exists():
+        logging.info("Saving to database")
+        proposals_v2 = adapter2.validate_json(proposal_v2_path.read_bytes())
+        mapping = adapter_meetings_map.validate_json(meeting_mapping_path.read_bytes())
+        save_to_db(db_path, proposals_v2, mapping, project_id)
 
 
 if __name__ == "__main__":
